Moved AccountSettingContent content struct to AccountSettingContent.swift file
