+++ Sources/App/Content/AccountSettingContent.swift
@@ -1,0 +1,8 @@
+import Vapor
+
+struct AccountSettingContent: Content {
+    let id: AccountSetting.ID?
+    let accountID: Account.ID?
+    let name: String
+    let value: String
+}
